+++ services/web/server/src/simcore_service_webserver/login/config.py
@@ -3,60 +3,19 @@
     - config-file schema
     - settings
 """
-from typing import Dict, Optional
+from typing import Dict
 
-import trafaret as T
 from aiohttp.web import Application
-from pydantic import BaseSettings
-
 from servicelib.aiohttp.application_keys import APP_CONFIG_KEY
 
 from ..email_config import CONFIG_SECTION_NAME as SMTP_SECTION
-from .cfg import DEFAULTS
+from ._schema import CONFIG_SECTION_NAME
 from .storage import AsyncpgStorage
-
-CONFIG_SECTION_NAME = "login"
-
-# TODO: merge with cfg.py
-schema = T.Dict(
-    {
-        T.Key("enabled", default=True, optional=True): T.Bool(),
-        T.Key(
-            "registration_confirmation_required",
-            default=DEFAULTS["REGISTRATION_CONFIRMATION_REQUIRED"],
-            optional=True,
-        ): T.Or(T.Bool, T.ToInt),
-        T.Key("registration_invitation_required", default=False, optional=True): T.Or(
-            T.Bool, T.ToInt
-        ),
-    }
-)
-
-
-class LoginSettings(BaseSettings):
-    enabled: bool = True
-    registration_confirmation_required: Optional[bool] = DEFAULTS[
-        "REGISTRATION_CONFIRMATION_REQUIRED"
-    ]
-    registration_invitation_required: Optional[bool] = False
-
-    class Config:
-        case_sensitive = False
-        env_prefix = "WEBSERVER_"
 
 
 def get_login_config(app: Application) -> Dict:
 
     cfg = app[APP_CONFIG_KEY].get(CONFIG_SECTION_NAME, {})
-    return cfg
-
-
-def assert_valid_config(app: Application) -> Dict:
-    """
-    raises pydantic.ValidationError if validation fails
-    """
-    cfg = get_login_config(app)
-    _settings = LoginSettings(**cfg)
     return cfg
