+++ services/web/server/src/simcore_service_webserver/login/module_setup.py
@@ -11,8 +11,8 @@ from servicelib.common_aiopg_utils import DSN
 from .._constants import APP_OPENAPI_SPECS_KEY, INDEX_RESOURCE_NAME
 from ..db_config import CONFIG_SECTION_NAME as DB_SECTION
 from .cfg import APP_LOGIN_CONFIG, cfg
-from .config import assert_valid_config, create_login_internal_config
 from .routes import create_routes
+from .settings import assert_valid_config, create_login_internal_config
 from .storage import AsyncpgStorage
 
 log = logging.getLogger(__name__)
